Week 4 is done

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -6,12 +6,14 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
+import org.webjars.NotFoundException;
 import ru.skypro.homework.dto.ad.AdDto;
 import ru.skypro.homework.dto.ad.AdsDto;
 import ru.skypro.homework.dto.ad.CreateOrUpdateAdDto;
 import ru.skypro.homework.dto.ad.ExtendedAdDto;
 import ru.skypro.homework.entity.Ad;
 import ru.skypro.homework.entity.User;
+import ru.skypro.homework.exception.AdNotFoundException;
 import ru.skypro.homework.mapper.AdMapper;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.UserRepository;
@@ -71,7 +73,7 @@ public class AdServiceImpl implements AdService {
 
     @Override
     public Ad getAd(Integer id) {
-        return adRepository.findById(id).orElseThrow();
+        return adRepository.findById(id).orElseThrow(() -> new AdNotFoundException(id));
     }
 
     @Override
@@ -97,20 +99,23 @@ public class AdServiceImpl implements AdService {
 
     @Override
     public byte[] updateAdImage(Integer id, MultipartFile image, Authentication authentication) {
-
-            try {
-                Ad ad = getAd(id);
-                ad = uploadImage(ad, image);
-                return Files.readAllBytes(Path.of(ad.getImage()));
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
+        try {
+            Ad ad = getAd(id);
+            ad = uploadImage(ad, image);
+            return Files.readAllBytes(Path.of(ad.getImage()));
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
 
     }
 
     @Override
-    public void deleteAd(Integer id, Authentication authentication) {
-        adRepository.delete(getAd(id));
+    public void deleteAd(Integer id, Authentication authentication) throws NotFoundException {
+        if (adRepository.existsById(id)) {
+            adRepository.delete(getAd(id));
+        } else {
+            throw new NotFoundException("Ad is not found");
+        }
     }
